Rename GetSetReflectBaseTest to ReflectGetterSetterBaseTest.
